Audio, Display, Eeprom, Tsi an UI updates - Audio   - DevBeeper     * Update stop() to disable toggle.     * Override Toggle on() and off() to also check for pin being set.   - DevAudio     * Remove delay code that was commented since I'm pretty sure it doesn't help       with speaker pop. - Display   * Fix overflow condition in numberTypeValues() when display type is hex and     max is temporarily set to 65536.  Use uint32_t instead of uint16_t. - Eeprom   * Add additional Tsi configuration options for NSCN, PS, REFCHRG and EXTCHRG.     Use struct for setting and getting. - Tsi   * Use tsi0_isr() to catch possible error (electrode short) or overrun and     add valid() method that returns false if either occur.   * Add ability to configure/calibrate using NSCN, PS, REFCHRG, EXTCHRG and     Threshold.   * Add touched() method that returns true if read() is >= Threshold.   * Add check when acquiring a channel to stop(), enable pin and start()     only if pin not already enabled.   * Add static methods to get max register values for NSCN, PS, REFCHRG and     EXTCHRG and their actual values.   * Add reasonable defaults so that calibration can be reset. - UI   * Add checks for Tsi invalidity based on above error/overruns.   - Alarm     * Add a touch disabled state if touch is triggered too quickly upon       alarming.  It's likely that the threshold is set too low.   * Update places where _blink.reset() is used to use a constant for setting     on one for done that are defined in the UISetState base class.   * Add break to a couple of default cases in switch statements that were     missed.   - Clock     * Only call _alarm.snooze() in uisUpdate() if, in addition to the alarm       being inProgress(), is not currently snoozing.  This was forcing a       snooze on an encoder turn, thus resetting the snooze start time and was       also preventing the cycling through night light and palettes.   - Touch     * Add ability to configure NSCN (Number of Consecutive Scans per Electrode),       PS (Electrode Oscillator PreScaler), REFCHRG (Reference Oscillator Charge       Current) and EXTCHRG (Electrode Oscillator Charge Current).     * Make them only configurable in an "advanced" state requiring a       PS_ALT_STATE press.     * Work in tandem with Tsi module to calibrate.

+++ display.cpp
@@ -77,7 +77,7 @@ void Seg7Display::setColon(df_t flags)
     _positions[2] = (flags & DF_COLON) ? _colon : 0x00;
 }
 
-void Seg7Display::numberTypeValues(uint16_t & max, uint16_t & num, uint16_t & den, nd_e nd, df_t flags)
+void Seg7Display::numberTypeValues(uint32_t & max, uint32_t & num, uint32_t & den, nd_e nd, df_t flags)
 {
     if (flags & DF_OCT)
     {
@@ -164,7 +164,7 @@ void Seg7Display::setInteger(int32_t integer, dp_e dp, nd_e nd, df_t flags)
 
 
     nd_e ntv_nd = (dp + nd) > ND_4 ? (nd_e)(nd - dp) : nd;
-    uint16_t max, num, den;
+    uint32_t max, num, den;
     numberTypeValues(max, num, den, ntv_nd, flags);
 
     n = validateValue(n, 0, max);
